\name{pheatmap_schemes}
\alias{pheatmap_schemes}
\title{List built-in pheatmapColor schemes}
\usage{
pheatmap_schemes()
}
\value{
A character vector of built-in scheme names.
}
\description{
List the built-in schemes that can be supplied to \code{scheme} in
\code{pheatmap_bio()} and \code{bio_annotation_colors()}.
}
\examples{
pheatmap_schemes()
}
